+++ Sources/SuperwallKit/Paywall/Presentation/Audience Logic/Expression Evaluator/EvaluationContext.swift
@@ -17,6 +17,14 @@ final class EvaluationContext: HostContext {
   }
 
   func computedProperty(name: String, args: String, callback: ResultCallback) {
+    evaluateProperty(name: name, args: args, callback: callback)
+  }
+
+  func deviceProperty(name: String, args: String, callback: ResultCallback) {
+    evaluateProperty(name: name, args: args, callback: callback)
+  }
+
+  private func evaluateProperty(name: String, args: String, callback: ResultCallback) {
     Task {
       guard
         let type = ComputedPropertyRequestType.allCases.first(where: { $0.description == name })
@@ -96,10 +104,6 @@ final class EvaluationContext: HostContext {
       }
       callback.onResult(result: jsonString)
     }
-  }
-
-  func deviceProperty(name: String, args: String, callback: ResultCallback) {
-    callback.onResult(result: "")
   }
 }