+++ Sources/SuperwallKit/StoreKit/Products/Receipt Manager/Receipt Manager/ReceiptManager.swift
@@ -79,6 +79,9 @@ actor ReceiptManager {
         case .verified(let transaction),
           .unverified(let transaction, _):
           Self.appTransactionId = transaction.appTransactionID
+          if Superwall.isInitialized {
+            Superwall.shared.dequeueIntegrationAttributes()
+          }
         }
       }
     }